Pull in cargo-xbuild as a dep
Recently encountered issue where a newer version of cargo-xbuild
caused cargo-n64 to not function. Pull the application in as a
dependency instead, so we have greater control over the processes
involved in building an N64 ROM.

The only code change is to invoke cargo-n64 itself instead of
cargo-xbuild. cargo_n64::run is modified to handle this kind of
invocation.