ROM padding corrections
CRCs ended up being incorrect here, since CRCs are calculated before
any of the total ROM padding is done. Instead, pad the program vector
first, so the correct the CRCs are calculated. Then combine everything
into a ROM and perform the final padding.

Also changes the padding size to be either a power of 2, or a multiple
of 4 MiB, whichever is smallest.