Section & ROM padding updates (#23)
* Section & ROM padding updates

Align all sections to 16 bytes. Pad the output ROM with 0xFF bytes until
the next power of 2.

* ROM padding corrections

CRCs ended up being incorrect here, since CRCs are calculated before
any of the total ROM padding is done. Instead, pad the program vector
first, so the correct the CRCs are calculated. Then combine everything
into a ROM and perform the final padding.

Also changes the padding size to be either a power of 2, or a multiple
of 4 MiB, whichever is smallest.